ESPWiFiCheerlightsNeopixels.ino
Takes the esp-cheer-lights WiFi code and merges it with the Tweetmas-Tree Arduino Ethernet code (see below) to produce a varying Cheerlight connected display suitable for a my (small) Christmas tree. Note that some of the routines from the Tweetmas-Tree code are left in so that the user can add in more routines as needed. Tweemas-Tree code can be found here: 
https://codebender.cc/sketch:69530#Arduino%20Tweetmas%20Tree%20Code%20by%20totof60%20NeoPixels%20Ethernet%20Cheerlights.ino